datepickerExample on instrumentation dashboard && new build
